requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

# ─── Medical LLM (GGUF inference — low VRAM) ─────────────────────────────────
llama-cpp-python==0.3.4

# ─── Optional: Redis Cache Backend ───────────────────────────────────────────
redis==5.2.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0

# ─── HTTP Client (for testing FastAPI) ────────────────────────────────────────
httpx==0.28.1
